requirements.txt
langchain
langchain_core
python-dotenv
langchain_groq
